offline-phase: lowgear: shared-random: Generate shared random values
These shared random values are used to construct shared bits and inverse
tuples

+++ mp-spdz-rs/src/fhe/params.rs
@@ -57,13 +57,13 @@ impl<C: CurveGroup> BGVParams<C> {
     }
 
     /// Get the number of plaintext slots the given parameters support
-    pub fn plaintext_slots(&self) -> u32 {
-        self.as_ref().n_plaintext_slots()
+    pub fn plaintext_slots(&self) -> usize {
+        self.as_ref().n_plaintext_slots() as usize
     }
 
     /// Get the number of ciphertexts that may be proven together
     pub fn ciphertext_pok_batch_size(&self) -> usize {
-        (self.plaintext_slots() as usize) * (DEFAULT_DROWN_SEC as usize)
+        self.plaintext_slots() * (DEFAULT_DROWN_SEC as usize)
     }
 }
